requirements.txt
pyrofork
pydantic
pydantic-settings
google-genai
# tmdbsimple removed
python-dotenv
aiofiles
tenacity
tgcrypto
